fix cyclic dep, create with plugin, update with key
